Use the existing setUpDB methods
Also, perform normal installs on all plugins and register them correctly during the base DB install.
